refactor(core): move formatting logic to FormatService
