+++ src/lib/commands/format.ts
@@ -1,23 +1,15 @@
 import { InteractionReplyOptions } from 'discord.js';
 import { DITypes } from '../container/container';
-import {
-  languageMappings,
-  languageNameMappings,
-} from '../formatters/FormatterMappings';
 import { ICommand, COMMAND_TYPE } from '../interfaces/ICommand';
-import { IDetector } from '../interfaces/IDetector';
-import { IParser } from '../interfaces/IParser';
-import { reformat } from '../util/reformatter';
-import { checkIfLanguageSupported, commentify } from '../util/utils';
+import { FormatService } from '../service/FormatService';
 
 const FormatCommand: ICommand<COMMAND_TYPE.SLASH> = {
   name: 'Format',
   description: '',
   type: COMMAND_TYPE.SLASH,
   async execute(interaction, container) {
-    // Get parser
-    const parser = container.getByKey<IParser>(DITypes.parser);
-    const detector = container.getByKey<IDetector>(DITypes.detector);
+    // Get service
+    const service = container.getByKey<FormatService>(DITypes.formatService);
     const baseReply: InteractionReplyOptions = { ephemeral: true };
     await interaction.deferReply(baseReply);
     // Parse package.json first
@@ -25,87 +17,8 @@ const FormatCommand: ICommand<COMMAND_TYPE.SLASH> = {
       const message = await interaction.channel.messages.fetch(
         interaction.targetId
       );
-      const blocks = parser.parseMessage(message.toString());
-      if (blocks.length === 0)
-        return interaction.editReply({
-          ...baseReply,
-          content: "Couldn't find any code blocks in this message.",
-        });
-      // Find formatter and try to format
-      const formattedBlocks = await Promise.all(
-        blocks.map(async (block) => {
-          let formattedBlock;
-          let detectedLanguageKey;
-          if (block.languageKey) {
-            try {
-              formattedBlock = await languageMappings[block.languageKey].format(
-                block.content
-              );
-            } catch (err) {
-              const comment = commentify(
-                `Couldn't format this ${
-                  languageNameMappings[block.languageKey]
-                } snippet. Perhaps there's a syntax error?`,
-                block.languageKey
-              );
-              formattedBlock = `${comment}\n${block.content}`;
-            }
-          } else {
-            // Attempt detecting the language
-            const { langKey, fullLangName } = await detector.detect(
-              block.content
-            );
-            detectedLanguageKey = langKey;
-            if (detectedLanguageKey) {
-              // One last attempt at formatting
-              // Check if the language is supported
-              const isSupported = checkIfLanguageSupported(detectedLanguageKey);
-              if (isSupported) {
-                try {
-                  formattedBlock = await languageMappings[
-                    detectedLanguageKey
-                  ].format(block.content);
-                } catch (err) {
-                  const comment = commentify(
-                    `Couldn't format this snippet. Perhaps there's a syntax error, or maybe the detector made a mistake?${
-                      fullLangName
-                        ? ` Detected language: ${
-                            languageNameMappings[detectedLanguageKey]
-                              ? languageNameMappings[detectedLanguageKey]
-                              : fullLangName
-                          }`
-                        : ''
-                    }`
-                  );
-                  // Formatting failed, set detectedLanguageKey to null.
-                  detectedLanguageKey = null;
-                  formattedBlock = `${comment}\n${block.content}`;
-                }
-              }
-            } else {
-              const comment = commentify(
-                `Couldn't find a compatible formatter for this code block. ${
-                  fullLangName
-                    ? `detected language: ${
-                        languageNameMappings[detectedLanguageKey]
-                          ? languageNameMappings[detectedLanguageKey]
-                          : fullLangName
-                      }`
-                    : "Couldn't detect a language."
-                }`
-              );
-              formattedBlock = `${comment}\n${block.content}`;
-            }
-          }
-          let reformatLangKey = block.languageKey
-            ? block.languageKey
-            : detectedLanguageKey;
-          reformatLangKey ??= '';
-          return reformat(formattedBlock, reformatLangKey);
-        })
-      );
-      const reply = formattedBlocks.join('');
-      interaction.editReply({ ...baseReply, content: reply });
+      const formatted = await service.format(message.toString());
+      interaction.editReply({ ...baseReply, content: formatted });
     } catch (err) {
       interaction.editReply({
         ...baseReply,
